# api/openapi.yaml
openapi: 3.1.0
info:
  title: BookLab API
  description: Self-hosted resource booking system with hourly post-use billing via Stripe.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: /api
    description: Relative to the running server
...
paths:
...
  /bookings:
    post:
      tags: [public]
      summary: Create a booking
      description: Creates a confirmed booking and returns a Stripe SetupIntent client_secret for card collection.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Time slot conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /bookings/{token}/cancel:
    post:
      tags: [public]
      summary: Cancel booking via magic token
      operationId: cancelBooking
      parameters:
        - $ref: "#/components/parameters/CancelToken"
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingPublic"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Booking cannot be cancelled (already cancelled or charged)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: booklab_session
  parameters:
    CancelToken:
      name: token
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    BookingStatus:
      type: string
      enum: [confirmed, cancelled, completed, charged]
...
    CreateBookingRequest:
      type: object
      required: [name, email, start_time, end_time]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
    CreateBookingResponse:
      type: object
      required: [booking, setup_intent_client_secret]
      properties:
        booking:
          $ref: "#/components/schemas/BookingPublic"
        setup_intent_client_secret:
          type: string
    BookingPublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/BookingStatus"
        cancel_token:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...