# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// Manually maintained equivalent — regenerate with `sqlc generate`.

package db

import (
	"time"

	"github.com/google/uuid"
)

type BookingStatus string

const (
	BookingStatusConfirmed  BookingStatus = "confirmed"
	BookingStatusCancelled  BookingStatus = "cancelled"
	BookingStatusCompleted  BookingStatus = "completed"
	BookingStatusCharged    BookingStatus = "charged"
)

type Booking struct {
	ID                     int32         `json:"id"`
	Name                   string        `json:"name"`
	Email                  string        `json:"email"`
	StartTime              time.Time     `json:"start_time"`
	EndTime                time.Time     `json:"end_time"`
	Status                 BookingStatus `json:"status"`
	CancelToken            uuid.UUID     `json:"cancel_token"`
	StripeSetupIntentID    *string       `json:"stripe_setup_intent_id"`
	StripePaymentMethodID  *string       `json:"stripe_payment_method_id"`
	StripePaymentIntentID  *string       `json:"stripe_payment_intent_id"`
	AmountCents            *int32        `json:"amount_cents"`
	ReminderSent           bool          `json:"reminder_sent"`
	CreatedAt              time.Time     `json:"created_at"`
	UpdatedAt              time.Time     `json:"updated_at"`
}

type Closure struct {
	ID        int32     `json:"id"`
	StartDate time.Time `json:"start_date"`
	EndDate   time.Time `json:"end_date"`
	Reason    *string   `json:"reason"`
	CreatedAt time.Time `json:"created_at"`
}

type Settings struct {
	ID                   int32     `json:"id"`
	ResourceName         string    `json:"resource_name"`
	HourlyRateCents      int32     `json:"hourly_rate_cents"`
	Currency             string    `json:"currency"`
	Timezone             string    `json:"timezone"`
	BookableStart        time.Time `json:"bookable_start"`
	BookableEnd          time.Time `json:"bookable_end"`
	MinHours             int32     `json:"min_hours"`
	MaxHours             int32     `json:"max_hours"`
	ReminderHoursBefore  int32     `json:"reminder_hours_before"`
}

type AdminUser struct {
	ID           int32     `json:"id"`
	Username     string    `json:"username"`
	PasswordHash string    `json:"password_hash"`
	CreatedAt    time.Time `json:"created_at"`
}